refactor(gui-v2): combine composables
