+++ packages/nc-gui-v2/components/dlg/TableCreate.vue
@@ -2,7 +2,7 @@
 import type { ComponentPublicInstance } from '@vue/runtime-core'
 import { Form } from 'ant-design-vue'
 import { useToast } from 'vue-toastification'
-import { onMounted, useProject, useTableCreate, useTabs } from '#imports'
+import { onMounted, useProject, useTable, useTabs } from '#imports'
 import { validateTableName } from '~/utils/validation'
 import { TabType } from '~/composables'
 
@@ -29,7 +29,7 @@ const { addTab } = useTabs()
 
 const { loadTables } = useProject()
 
-const { table, createTable, generateUniqueTitle, tables, project } = useTableCreate(async (table) => {
+const { table, createTable, generateUniqueTitle, tables, project } = useTable(async (table) => {
   await loadTables()
 
   addTab({
