+++ packages/nc-gui-v2/components/dlg/TableRename.vue
@@ -3,7 +3,7 @@ import { watchEffect } from '@vue/runtime-core'
 import { Form } from 'ant-design-vue'
 import type { TableType } from 'nocodb-sdk'
 import { useToast } from 'vue-toastification'
-import { useProject, useTableCreate, useTabs } from '#imports'
+import { useProject, useTabs } from '#imports'
 import { extractSdkResponseErrorMsg } from '~/utils/errorUtils'
 import { validateTableName } from '~/utils/validation'
 import { useNuxtApp } from '#app'
