+++ packages/nc-gui-v2/composables/index.ts
@@ -10,7 +10,7 @@ export * from './useHasMany'
 export * from './useManyToMany'
 export * from './useMetas'
 export * from './useProject'
-export * from './useTableCreate'
+export * from './useTable'
 export * from './useTabs'
 export * from './useViewColumns'
 export * from './useViewData'
@@ -18,6 +18,5 @@ export * from './useViewFilters'
 export * from './useViews'
 export * from './useViewSorts'
 export * from './useVirtualCell'
-export * from './useDeleteTable'
 export * from './useColumnCreateStore'
 export * from './useSmartsheetStore'
